requirements.txt
pymongo==4.6.1
openai>=1.40
python-dotenv==1.0.1
httpx<0.28
httpcore<1.0